Revert "Add single map"
