+++ src/app/[...slug]/page.jsx
@@ -1,6 +1,21 @@
 import { notFound } from 'next/navigation';
+import { Hero } from '../../components/Hero.jsx';
 import { getPageFromSlug } from '../../utils/content.js';
-import { getComponentForSection } from '../../components/index.jsx';
+import NavLinks from '../../components/NavLinks.jsx';
+import Playground from '../../components/Playground.jsx';
+import ContentWithImage from '../../components/ContentWithImage.jsx';
+import Carousel from '../../components/Carousel.jsx';
+import Footer from '../../components/Footer.jsx';
+import { Fallback } from '../../components/Fallback.jsx';
+
+const componentMap = {
+  hero: Hero,
+  navigation: NavLinks,
+  playground: Playground,
+  contentWithImage: ContentWithImage,
+  carousel: Carousel,
+  footer: Footer,
+};
 
 export default async function ComposablePage({ params }) {
   const { slug } = params;
@@ -17,7 +32,7 @@ export default async function ComposablePage({ params }) {
     return (
       <div data-sb-object-id={page.id}>
         {(page.sections || []).map((section, idx) => {
-          const Component = getComponentForSection(section);
+          const Component = componentMap[section.type] || Fallback;
           return <Component key={idx} {...section} />;
         })}
       </div>
