+++ src/app/layout.jsx
@@ -3,7 +3,9 @@ import '../../styles/globals.css';
 export default async function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body>
+        {children}
+      </body>
     </html>
   );
 }
