+++ src/app/page.jsx
@@ -1,6 +1,20 @@
 import { notFound } from 'next/navigation';
+import { Hero } from '../components/Hero.jsx';
 import { getPageFromSlug } from '../utils/content.js';
-import { getComponentForSection } from '../components';
+import NavLinks from '../components/NavLinks.jsx';
+import Playground from '../components/Playground.jsx';
+import ContentWithImage from '../components/ContentWithImage.jsx';
+import Carousel from '../components/Carousel.jsx';
+import Footer from '../components/Footer.jsx';
+
+const componentMap = {
+  hero: Hero,
+  navigation: NavLinks,
+  playground: Playground,
+  contentWithImage: ContentWithImage,
+  carousel: Carousel,
+  footer: Footer,
+};
 
 export default async function ComposablePage() {
   try {
@@ -13,7 +27,7 @@ export default async function ComposablePage() {
     return (
       <div data-sb-object-id={page.id}>
         {(page.sections || []).map((section, idx) => {
-          const Component = getComponentForSection(section);
+          const Component = componentMap[section.type] || Fallback;
           return <Component key={idx} {...section} />;
         })}
       </div>
