+++ src/components/Fallback.jsx
@@ -1,0 +1,3 @@
+export const Fallback = (props) => {
+  return <p>{JSON.stringify(props, null, 4)}</p>;
+};
